fix(target_actor): prevent inconsistent adds

+++ front/targetticket.form.php
@@ -41,31 +41,51 @@ $targetticket = new PluginFormcreatorTargetTicket();
 
 // Edit an existing target ticket
 if (isset($_POST['update'])) {
-   $targetticket->update($_POST);
+   if (!$targetticket->canUpdateItem()) {
+      Session::addMessageAfterRedirect(__('No right to update this item.', 'formcreator'), false, ERROR);
+   } else {
+      $targetticket->update($_POST);
+   }
    Html::back();
 
 } else if (isset($_POST['actor_role'])) {
-   $id          = (int) $_POST['id'];
-   $actor_value = $_POST['actor_value_' . $_POST['actor_type']] ?? null;
-   $use_notification = ($_POST['use_notification'] == 0) ? 0 : 1;
-   $targetTicket_actor = new PluginFormcreatorTarget_Actor();
-   $targetTicket_actor->add([
-      'itemtype'         => $targetticket->getType(),
-      'items_id'         => $id,
-      'actor_role'       => $_POST['actor_role'],
-      'actor_type'       => $_POST['actor_type'],
-      'actor_value'      => $actor_value,
-      'use_notification' => $use_notification,
-   ]);
+   $id = (int) $_POST['id'];
+   $targetticket->getFromDB($id);
+   if (!$targetticket->canUpdateItem()) {
+      Session::addMessageAfterRedirect(__('No right to update this item.', 'formcreator'), false, ERROR);
+   } else {
+      $actor_value = $_POST['actor_value_' . $_POST['actor_type']] ?? null;
+      $use_notification = ($_POST['use_notification'] == 0) ? 0 : 1;
+      $targetTicket_actor = new PluginFormcreatorTarget_Actor();
+      $targetTicket_actor->add([
+         'itemtype'         => $targetticket->getType(),
+         'items_id'         => $id,
+         'actor_role'       => $_POST['actor_role'],
+         'actor_type'       => $_POST['actor_type'],
+         'actor_value'      => $actor_value,
+         'use_notification' => $use_notification,
+      ]);
+   }
    Html::back();
 
 } else if (isset($_GET['delete_actor'])) {
-   $targetTicket_actor = new PluginFormcreatorTarget_Actor();
-   $targetTicket_actor->delete([
-      'itemtype' => $targetticket->getType(),
-      'items_id' => $id,
-      'id'       => (int) $_GET['delete_actor']
-   ]);
+   $requiredKeys = ['id'];
+   if (count(array_intersect(array_keys($_GET), $requiredKeys)) < count($requiredKeys)) {
+      Session::addMessageAfterRedirect(__('Bad request while deleting an actor.', 'formcreator'), false, ERROR);
+      Html::back();
+   }
+   $id = (int) $_GET['id'];
+   $targetticket->getFromDB($id);
+   if (!$targetticket->canUpdateItem()) {
+      Session::addMessageAfterRedirect(__('No right to update this item.', 'formcreator'), false, ERROR);
+   } else {
+      $targetTicket_actor = new PluginFormcreatorTarget_Actor();
+      $targetTicket_actor->delete([
+         'itemtype' => $targetticket->getType(),
+         'items_id' => $id,
+         'id'       => (int) $_GET['delete_actor']
+      ]);
+   }
    Html::back();
 
    // Show target ticket form
